chore: refactored check answers controller & case api
